// src/types.tsx
export interface Course {
    id: number;
    slug: string;
    image: string;
    category_id: number;
    category_name: string;
    course_type: string;
    level: string;
    sequential_access?: boolean;
    certificate_enabled?: boolean;
    certificate_fee?: number | null;
    ratings: number;
    review_count?: number;
    average_rating: number;
    title: string;
    short_description: string;
    author_id?: number;
    author_name: string;
    author_slug?: string;
    price: number;
    // discounted_price?: number;
    discount_price?: number;
    discount_percentage: number;
    total_tax_percentage?: number;
    tax_amount?: number;
    is_wishlisted: boolean;
    is_enrolled: boolean;
    // for my learning
    enrolled_at?: string;
    total_chapters?: number;
    completed_chapters?: number;
    total_curriculum_items?: number;
    completed_curriculum_items?: number;
    progress_percentage?: number;
    progress_status?: "all" | "in_progress" | "completed";
  }